Added in 2nd report

+++ WALE.ProcessFile.Models/NaldDataPoint.cs
@@ -3,7 +3,7 @@ namespace WALE.ProcessFile.Models;
 public class NaldDataPoint
 {
     public string? PointName { get; set; }
-    public int PointId { get; init; }
+    public long PointId { get; init; }
     public string? Ngr1 { get; set; }
     public string? Ngr1Cartesian { get; set; }
